requirements.txt
Django==1.8.5
django-crontab==0.7.0
django-haystack==2.4.1
django-widget-tweaks==1.4.1
elasticsearch==2.1.0
gunicorn==19.3.0
Pillow==3.0.0
urllib3==1.12
wheel==0.24.0